refactor(GAME): Reorganize frontend game files
