+++ front/src/components/play/CanvasUtils.tsx
@@ -1,6 +1,6 @@
-import { IGameInfos, IPlayer, IPosition } from './Game';
+import { IGameInfos, IGameState, IPlayer, IPosition } from './GameInterfaces';
 
-export function drawRect(
+function drawRect(
 	ctx: any,
 	color: string,
 	x: number,
@@ -16,7 +16,7 @@ export function drawRect(
 	ctx.fill();
 }
 
-export function drawPaddle(
+function drawPaddle(
 	ctx: any,
 	player: IPlayer,
 	color: string,
@@ -42,7 +42,7 @@ export function drawPaddle(
 	);
 }
 
-export function drawBall(
+function drawBall(
 	ctx: any,
 	color: string,
 	ball: IPosition,
@@ -52,4 +52,22 @@ export function drawBall(
 	ctx.beginPath();
 	ctx.arc(ball.x, ball.y, gameInfos.ballRadius, 0, 2 * Math.PI);
 	ctx.fill();
+}
+
+export function drawState(state: IGameState, canvasRef: any) {
+	const canvas: any = canvasRef.current;
+	const ctx = canvas.getContext('2d');
+
+	ctx.canvas.width = state.gameInfos.originalWidth;
+	ctx.canvas.height = state.gameInfos.originalHeight;
+
+	ctx.fillStyle = '#d9d9d9';
+	ctx.beginPath();
+	ctx.rect(0, 0, canvas.width, canvas.height);
+	ctx.fill();
+
+	drawBall(ctx, '#000000', state.ball, state.gameInfos);
+
+	drawPaddle(ctx, state.player1, '#ffb800', state.gameInfos);
+	drawPaddle(ctx, state.player2, '#17c0e9', state.gameInfos);
 }
